+++ src/sap_knowledge/knowledge/__init__.py
@@ -2,7 +2,7 @@
 
 from sap_knowledge.knowledge.chunking import CharacterChunker
 from sap_knowledge.knowledge.models import Citation, KnowledgeChunk, KnowledgeDocument
-from sap_knowledge.knowledge.recipes import FieldMapping, KnowledgeRecipe
+from sap_knowledge.knowledge.recipes import FieldMapping, KnowledgeRecipe, MetadataMapping
 from sap_knowledge.knowledge.rendering import KnowledgeRenderer, document_id_for
 
 __all__ = [
@@ -13,5 +13,6 @@ __all__ = [
     "KnowledgeDocument",
     "KnowledgeRecipe",
     "KnowledgeRenderer",
+    "MetadataMapping",
     "document_id_for",
 ]